12_Polymorphism lecture notes cont.

+++ module-1/12_Polymorphism/lecture/src/main/java/com/techelevator/farm/models/Sellable.java
@@ -1,8 +1,10 @@
 package com.techelevator.farm.models;
 
+import java.math.BigDecimal;
+
 public interface Sellable
 {
     String getName();
-    int getPrice();
+    BigDecimal getPrice();
 
 }
